Make product.py and category.py, tests. Delete task.py and test. Remake import
