+++ src/utils.py
@@ -1,7 +1,8 @@
 import json
 import os
 
-from src.task import Category, Product
+from src.product import Product
+from src.category import Category
 
 
 def read_json(path: str) -> dict:
