+++ tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
-from src.task import Category, Product
+from src.product import Product
+from src.category import Category
 
 
 @pytest.fixture
