+++ tests/test_category.py
@@ -1,10 +1,3 @@
-def test_product_init(product):
-    assert product.name == "Iphone 15"
-    assert product.description == "512GB, Gray space"
-    assert product.price == 210000.0
-    assert product.quantity == 8
-
-
 def test_category_init(first_category, second_category):
     assert first_category.name == "Смартфоны"
     assert first_category.description == (
