+++ tests/test_product.py
@@ -1,0 +1,5 @@
+def test_product_init(product):
+    assert product.name == "Iphone 15"
+    assert product.description == "512GB, Gray space"
+    assert product.price == 210000.0
+    assert product.quantity == 8
